+++ NEWS.md
@@ -1,4 +1,11 @@
-# simmer 4.4.5.1
+# simmer 4.4.5.2
+
+## New features
+
+- Add support for activity tags via `tag` argument (as part of #287).
+- Add support for named rollbacks via tags. The `amount` argument is deprecated
+  in favor of the more generic name `target`. The former will still work, but
+  raises a warning (#287 addressing #197).
 
 ## Minor changes and fixes